zeta: Do not show usage for copilot/supermaven
